+++ addons/services/libs/chunkStreams/chunkTxDataStream.ts
@@ -1,4 +1,4 @@
-import { getByteRange } from '../../libs/byte-ranges/byteRanges'
+import { getByteRange } from '../byte-ranges/byteRanges'
 import { chunkStream } from './chunkStream'
 import { ReadableStream } from 'node:stream/web'
 import { SIG_CONFIG, SignatureConfig, byteArrayToLong } from './ANS-104-constants'
